Added deleteAdmin functionality

+++ src/main/java/com/cpsc4910/cpscbackend/Admin/AdminRepository.java
@@ -2,5 +2,5 @@ package com.cpsc4910.cpscbackend.Admin;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AdminRepository extends CrudRepository<Admin, Integer>{
+public interface AdminRepository extends CrudRepository<Admin, Long>{
 }
